web_ng/requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiodns==4.0.0
aiofiles==25.1.0
aiogram==2.15
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==4.0.0
attrs==22.1.0
azure-cognitiveservices-speech==1.47.0
azure-core==1.38.0
babel==2.18.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
blis==1.2.1
boto3==1.42.42
botocore==1.42.42
CacheControl==0.12.14
cachetools==6.2.6
catalogue==2.0.10
ccxt==4.5.36
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
coincurve==21.0.0
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confection==0.1.5
confluent-kafka==2.13.0
cryptography==46.0.4
cycler==0.12.1
cymem==2.0.13
dask==2022.10.2
DAWG-Python==0.7.2
DAWG2-Python==0.9.0
diskcache==5.6.3
dnspython==2.3.0
docopt==0.6.2
docutils==0.22.4
exceptiongroup==1.3.1
fastapi==0.128.5
fbmessenger==6.0.0
filelock==3.20.3
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.4.0
generativeai==0.0.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.0
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
llama_cpp_python==0.2.90
locket==1.0.0
Markdown==3.10.1
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
multidict==5.2.0
murmurhash==1.0.15
narwhals==2.16.0
networkx==2.6.3
nicegui==3.7.1
numpy==1.23.5
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-runtime-cu12==12.9.79
oauthlib==3.3.1
opt_einsum==3.4.0
orjson==3.11.7
packaging==20.9
pamqp==3.2.1
pandas==2.3.3
parsedatetime==2.6
partd==1.4.2
pillow==12.1.0
pluggy==1.6.0
portalocker==2.10.1
preshed==3.0.12
prompt-toolkit==3.0.28
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.23.3
psutil==7.2.2
psycopg2-binary==2.9.11
pvporcupine==3.0.1
pvrecorder==1.2.7
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydot==1.4.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.11.0
pykwalify==1.8.0
pymongo==4.3.3
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.6
pymorphy3-dicts-ru==2.4.417150.4580142
pynput==1.8.1
pyparsing==3.3.2
pyreadline3==3.5.4
PyRect==0.2.0
pyrsistent==0.20.0
python-crfsuite==0.9.12
python-dateutil==2.8.2
python-dotenv==1.2.1
python-engineio==4.13.0
python-multipart==0.0.22
python-socketio==5.16.0
pytz==2022.7.1
pywin32==311
PyYAML==6.0.3
questionary==1.10.0
randomname==0.1.5
rarfile==4.2
rasa==3.6.21
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rocketchat-API==1.30.0
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.15
s3transfer==0.16.0
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
skops==0.9.0
slack_sdk==3.39.0
smart_open==7.5.0
smmap==5.0.2
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.5.2
starlette==0.52.1
streamlit==1.54.0
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tenacity==9.1.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
terminaltables==3.1.10
thinc==8.3.4
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
twilio==8.2.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing-utils==0.1.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.5.3
weasel==0.4.3
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.1.5
WMI==1.5.1
wrapt==1.14.2
wsproto==1.3.2
yarl==1.22.0
